+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ dev
 ---
 - Added new example: the stochastic Lotka-Volterra model
 - Fix methods.bo.utils.minimize to be strictly within bounds
+- Added the MaxVar acquisition
 
 0.6.3 (2017-09-28)
 ------------------